+++ noodles-vcf/CHANGELOG.md
@@ -7,6 +7,8 @@
   * vcf/io/writer/record: Add validation for IDs, reference bases, alternate
     bases, and filters.
 
+  * vcf/io/writer/info/field/value/string: Percent-encode special characters.
+
 ## 0.52.0 - 2024-04-04
 
 ### Added